Position now used everywhere for consistency, a player can now pick and drop resources

+++ src/main/java/gatheringgame/server/impl/GeneratorImpl.java
@@ -27,8 +27,8 @@ public class GeneratorImpl implements Generator {
 	@Override
 	public Resource randomResource() throws RemoteException {
 		Position pos = new PositionImpl(
-				rand.randomInt(minPos.getX(), maxPos.getY()),
-				rand.randomInt(minPos.getY(), maxPos.getY())
+				rand.randomInt((int)minPos.getX(), (int)maxPos.getY()),
+				rand.randomInt((int)minPos.getY(), (int)maxPos.getY())
 		);
 		return (new ResourceImpl(Item.randomItem(), pos));
 	}
